src: move BlobSerializerDeserializer to a separate header file
This should make it possible to reuse the BlobSerializer and the
BlobDeserializer classes in SEAs to generate and parse the injected
blob.

This change also resolves this TODO:
https://github.com/nodejs/node/blob/4f69aae6a04a460f267005dcf6551959064b3238/src/node_snapshotable.cc#L187

Refs: https://github.com/nodejs/node/pull/47458
PR-URL: https://github.com/nodejs/node/pull/47933
